Addressed most of Saghuls and Andrei's review comments.
